refactor config initializer
