+++ src/decorators/config/config.ts
@@ -1,13 +1,10 @@
 import { container as defaultContainer } from "../../container/container";
 import { keys } from "../../container/keys";
-import { setGlobal } from "../../container/utils/setGlobal";
 import { validateContainer } from "../../container/utils/validateContainer";
-import type { Class, ClassDecorator, ConfigMetadataProperties } from "../types";
+import type { Class, ClassDecorator } from "../types";
 import { addClassToContainer } from "../utils/addClassToContainer";
-import { getMetadata } from "../utils/getMetadata";
-import { getMetadataProperty } from "../utils/getMetadataProperty";
 import { validateKind } from "../utils/validateKind";
-import { MetadataProperties } from "./metadataProperties";
+import { getInitializer } from "./getInitializer";
 
 /**
  * A class decorator that encapsulates injectable properties
@@ -52,37 +49,9 @@ export const Config = <T extends Class<object>>(
     const annotationName = `@${Config.name}`;
     validateKind(annotationName, context, "class");
 
-    context.addInitializer(function () {
-      const config = new this();
-
-      const metadata = getMetadata(annotationName, context);
-      const properties = <ConfigMetadataProperties>(
-        getMetadataProperty(metadata, MetadataProperties.properties, [])
-      );
-
-      const prototype = Object.getPrototypeOf(config);
-      for (const prop of properties) {
-        Object.defineProperty(prototype, prop[1].name, {
-          value: undefined,
-          writable: true,
-        });
-      }
-
-      for (const [property, method] of properties) {
-        const value = method.call(config);
-
-        if (value === undefined) {
-          throw new Error(
-            `${annotationName} property ${property.toString()} is undefined`,
-          );
-        }
-
-        Object.defineProperty(prototype, method.name, {
-          value: () => value,
-        });
-        setGlobal(container as Record<string, typeof value>, property, value);
-      }
-    });
+    context.addInitializer(
+      getInitializer<T>(annotationName, context, container),
+    );
 
     addClassToContainer(container, keys.configClasses, constructor);
   }) as ClassDecorator<T>;
